show failed command output when debugging is enabled
